[MH3-97655] - Increased the UDP Rx Queue Limit type from uint8_t to uint16_t

+++ tcpip/src/udp_private.h
@@ -244,7 +244,7 @@ typedef struct
     TCPIP_UDP_SIGNAL_FUNCTION sigHandler;  // socket event handler
     const void*     sigParam;               // event handler parameter
     uint16_t        sigMask;        // TCPIP_UDP_SIGNAL_TYPE: active events
-    uint8_t         rxQueueLimit;   // max number of RX packets that can be queued at a certain time
+    uint16_t         rxQueueLimit;   // max number of RX packets that can be queued at a certain time
     uint8_t         ttl;            // socket TTL value 
     uint8_t         padding[];      // padding; not used
